Adds Y origin offset and corrects all initial port positions accordingly (#3 and #4)
The first row of the bitmap and mesh tile is actually unused
and was not part of the height count.

So either we needed to add an additional row,
or we needed to offset the display (which I did).

I spotted this, because I observed that the wind mesh was not rendered on last row
(for instance, Geneva port had no wind).